Refactor legacy tests into focused files
