--- tests/Lokad.Onnx.Backend.Tests/Me5sModelTests.cs
@@ -1,11 +1,0 @@
-﻿namespace Lokad.Onnx.Backend.Tests;
-
-using Xunit;
-public class Me5sModelTests : Runtime
-{
-    static Me5sModelTests()
-    {
-        
-    }
-}
-
